Finish

+++ include/reader.hpp
@@ -10,6 +10,6 @@ class CSVReader {
   std::vector<double> stripLine_(std::string line);
 
  public:
-  std::pair<std::vector<std::vector<double>>, std::vector<double>> getData(
+  std::pair<std::vector<std::vector<double>>, std::vector<long>> getData(
       std::string filename, bool has_class_column, int skip_lines);
 };
